Repository Pattern Logic, repo user interfaces and imp classes.
A base repository was created that hosts the basic CRUD methods, both an interface in contracts and its implementation in repositories. Then, the repository for our entities in the interface and its implementation that inherits from the base repository, and the interface in case something more than the CRUD methods is required, were created.